Some tweaks in the graphics
The lane markings have been made smaller so they fit inside blocks of 24x24 pixels.

The grey shade of the bus stop sign base has been changed to match the one of overhead signs.

On the logos, the words "The" and "Race" have been moved slightly so both have the "e" aligned to 8x8 tiles.